requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
cookies==2.2.1
Flask==0.12.2
funcsigs==1.0.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mock==2.0.0
pbr==3.1.1
python-forecastio==1.3.5
requests==2.18.4
responses==0.8.1
six==1.11.0
slackclient==1.1.0
urllib3==1.22
websocket-client==0.44.0
Werkzeug==0.13
